Semicolon: Add basic custom header functionality.
I think the theme actually works better without one, but this seems to be a function that a lot of people are looking for (https://wordpress.org/support/topic/header-image-116, https://wordpress.org/support/topic/how-would-i-add-a-responsive-full-with-header-image), and it doesn't hurt to add it.

In the interests of keeping things as straightforward as possible, I'm using as simple an implementation of this functionality as possible.