Independent Publisher: Disable Gravatar Hovercards on the site logo.
A default avatar is displayed in the place of the site logo
when there is no Jetpack Site Logo set. This avatar triggers
a Gravatar Hovercard, which serves no purpose.                  

JavaScript code has been added to disable Gravatar Hovercards
from being triggered on the site logo links (also applies to
the master logo).