Pique: Code simplifying/restructuring.
First, I've removed the "split nav" functionality from the menu (which
split the nav in half to the left and right of the logo). I like this
from a design perspective, and it looks really nice with a circular logo,
but it didn't look as nice with a horizontal logo. Removing it means the
code is simplified, which is good. 

Then, I rescaled the logo a bit to better fit in the newly-available space.

Finally, I rewrote the sticky navigation code, which was a bit of a mess and
was split between the front-page and the other-pages. Now it's all in one place
(and one JS file!) which should make debugging issues less of a monumental
pain in the [curse-word]. Refactoring the code also gave me an opportunity to
correct the issue where, if you refreshed the page when you were scrolled down
past the navigation, the sticky menu would be toggled the wrong way. Not sure any
users would ever actually do this, but I do it a lot and it was making me crazy.

Yay!