Pique: Remove double 'read more' link from posts.
Turned out I wasn't paying attention and added this code in two
different places; whoops! I've ajusted the CSS so that the read more
links look different in the two different areas where they appear:
button-y on the main blog archive pages, but a small text link only
on the blog panel that (can) display on the front page.